+++ Assets/Resources/Prefabs/PlayableCharacter/Gnome Tinkerer.prefab
@@ -13,7 +13,6 @@ GameObject:
   - component: {fileID: 3768773383118627058}
   - component: {fileID: 4409285678151217457}
   - component: {fileID: 466390313952093619}
-  - component: {fileID: 5115566951842065350}
   - component: {fileID: 7241768347325068628}
   - component: {fileID: 8663408860301830518}
   - component: {fileID: 2930062219622115911}
@@ -22,6 +21,7 @@ GameObject:
   - component: {fileID: 3229203563632729664}
   - component: {fileID: -5837867781125322070}
   - component: {fileID: 235036723466926854}
+  - component: {fileID: 4609410276060826380}
   - component: {fileID: 2298111754257009617}
   m_Layer: 7
   m_Name: Gnome Tinkerer
@@ -98,9 +98,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bef60c6cb1a38164f8090dbbfa438fb5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  Health: 20
+  statusAnchor: {fileID: 0}
   _resetOnStart: 1
   _hitAudioClip: {fileID: 8300000, guid: 281dc59a06f5bdb42a28ad3f59532891, type: 3}
   _armorAudioClip: {fileID: 8300000, guid: 51a3f7c7491b09d48adeff0b2c4ef7e6, type: 3}
+  _armorSprite: {fileID: 1441622573, guid: 9030d8e5d3ed0bf449dcf92ef5501231, type: 3}
   _onPlayerDeathEvent: {fileID: 11400000, guid: e52a29c257b87af4499524eee03d0667, type: 2}
 --- !u!114 &466390313952093619
 MonoBehaviour:
@@ -120,35 +123,6 @@ MonoBehaviour:
   _LevelUpEffectGO: {fileID: 5822821294587056223, guid: d51a89497c2380a40a1a2c8798adbfa3, type: 3}
   _LevelUpAudioClip: {fileID: 8300000, guid: 0e566a8701662924eb677419db5d35f9, type: 3}
   _playerLevelUpEvent: {fileID: 11400000, guid: 9fdb7426a996d4d4d99b3c5b345ec99a, type: 2}
---- !u!114 &5115566951842065350
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 704638989827276549}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9fc7297102e5d0d439bf3f73f83b567b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  gameEvent: {fileID: 11400000, guid: 532e962d558155b458520d5ff1f64c1c, type: 2}
-  unityEvent:
-    m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 4409285678151217457}
-        m_TargetAssemblyTypeName: Assets.Scripts.Controller.Player.PlayerHealth,
-          Assembly-CSharp
-        m_MethodName: TakeDamage
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
 --- !u!114 &7241768347325068628
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -344,6 +318,34 @@ MonoBehaviour:
         m_TargetAssemblyTypeName: Assets.Scripts.Controller.Player.PlayerHealth,
           Assembly-CSharp
         m_MethodName: OnPlayerHeal
+        m_Mode: 0
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
+--- !u!114 &4609410276060826380
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 704638989827276549}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 031776349f7507e47b97fc6d49ce5e2a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gameEvent: {fileID: 11400000, guid: c3494f9e28b23974f9f9bf4ae624ecef, type: 2}
+  unityEvent:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 4409285678151217457}
+        m_TargetAssemblyTypeName: Assets.Scripts.Controller.BaseHealth, Assembly-CSharp
+        m_MethodName: TakeDamage
         m_Mode: 0
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -625,7 +627,7 @@ MonoBehaviour:
   - {x: 0.002414398, y: -0.4909395, z: 0}
   m_ShapePathHash: 906204786
   m_Mesh: {fileID: 0}
-  m_InstanceId: 69948
+  m_InstanceId: 66420
   m_LocalBounds:
     m_Center: {x: -0.034057915, y: -0.12057042, z: 0}
     m_Extent: {x: 0.3386956, y: 0.37248492, z: 0}
