+++ Assets/Scripts/Controller/BaseHealth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62b65e2826986dc4c976aaf0145caf68
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
